+++ tt/src/main/java/com/oddlabs/tt/form/MapcodeForm.java
@@ -92,7 +92,7 @@ public final class MapcodeForm extends Form {
     private final class RandButtonListener implements MouseClickListener {
         @Override
         public void mouseClicked(@NonNull MouseButton button, int x, int y, int clicks) {
-            Random random = new Random(LocalEventQueue.getQueue().getHighPrecisionManager().getTick() * LocalEventQueue.getQueue().getHighPrecisionManager().getTick());
+            Random random = new Random(LocalEventQueue.getQueue().getHighPrecisionManager().getTick() * (long) LocalEventQueue.getQueue().getHighPrecisionManager().getTick());
             random.nextInt();
             BigInteger rand_int = new BigInteger(60, random);
             String rand_string = RegistrationKey.createString(rand_int);
